backend/requirements.txt
coreapi==2.3.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pillow==10.3.0
pip-chill==1.0.3
